refactor(ChessPieceBase): 체스말들의 공통 메서드를 상속받는 추상 클래스 ChessPieceBase 생성
- 생성에 따른 체스 말 객체 변화
