--- src/main/java/piece/Rook.java
@@ -1,33 +1,0 @@
-package piece;
-
-import coordinate.Coordinate;
-import java.util.List;
-import strategy.RookStrategy;
-
-public class Rook implements Piece {
-
-    private final Color color;
-
-    public Rook(Color color) {
-        this.color = color;
-    }
-
-    @Override
-    public List<Integer> getDirection(Coordinate coordinate, Coordinate nextCoordinate, boolean isAttack) {
-        int rowDifference = coordinate.checkRow(nextCoordinate);
-        int columnDifference = coordinate.checkColumn(nextCoordinate);
-
-        RookStrategy rookStrategy = RookStrategy.getMoveStrategy(rowDifference, columnDifference);
-        return rookStrategy.getDirection();
-    }
-
-    @Override
-    public boolean isSameColor(Color color) {
-        return this.color == color;
-    }
-
-    @Override
-    public boolean isBlack() {
-        return color == Color.BLACK;
-    }
-}
